% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{fit.submodel}
\alias{fit.submodel}
\title{Compute the fit of a submodel}
\usage{
fit.submodel(x, sigma2, mu, chosen, xt, yt, logistic)
}
\arguments{
\item{x}{Design matrix.}

\item{sigma2}{Residual variance (1 for logistic regression).}

\item{mu}{Matrix of fitted values for the full model.}

\item{chosen}{Vector of indices of the columns of \code{x} in the current submodel.}

\item{xt}{Design matrix for test data.}

\item{yt}{Outcome variable for test data.}

\item{logistic}{Set to \code{TRUE} for a logistic regression model.}
}
\description{
Compute the fit of a submodel
}
\keyword{internal}
