% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{get.coefficients}
\alias{get.coefficients}
\title{Returns the posterior means of the regression coefficients}
\usage{
get.coefficients(samples, coeff.names)
}
\arguments{
\item{samples}{An object of class \code{stanfit}.}

\item{coeff.names}{Vector of names for the coefficients.}
}
\value{
A list with two named elements: \var{unpenalized} for the posterior means
of the unpenalized covariates, and \var{penalized} for the posterior means
of the penalized predictors (which can be \code{NULL} for baseline models).
}
\description{
Returns the posterior means of the regression coefficients
}
