% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsstan-package.R
\docType{package}
\name{hsstan-package}
\alias{hsstan-package}
\title{The 'hsstan' package.}
\description{
Regression models with hierarchical shrinkage priors
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
