% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{hsstan}
\alias{hsstan}
\title{Fit a plain or cross-validated Stan model}
\usage{
hsstan(x, covs.model, penalized = NULL, family = gaussian,
  folds = NULL, iter = ifelse(is.null(folds), 2000, 1000),
  warmup = iter/2, scale.u = 2, regularized = TRUE,
  nu = ifelse(regularized, 1, 3), par.ratio = 0.05, global.df = 1,
  slab.scale = 2, slab.df = 4, qr = TRUE, seed = 123,
  adapt.delta = NULL, ...)
}
\arguments{
\item{x}{Data frame containing outcome, covariates and penalized predictors.
Continuous predictors and outcome variable should be standardized
before fitting the models as priors assume them to have mean zero and
have the same scale.}

\item{covs.model}{Formula containing the unpenalized covariates.}

\item{penalized}{Names of the variables to be used as penalized predictors.
If \code{NULL} or an empty vector, a model with only unpenalized
covariates is fitted.}

\item{family}{Type of model fitted: either \code{gaussian()} for linear
regression (default) or \code{binomial()} for logistic regression.}

\item{folds}{Integer vector with one element per observation indicating the
cross-validation fold in which the observation should be withdrawn.
If \code{NULL} (default), no cross-validation is performed.}

\item{iter}{Total number of iterations in each chain, including warmup
(by default, 1000 iterations for cross-validation and 2000 otherwise).}

\item{warmup}{Number of warmup iterations per chain (by default, half the
total number of iterations).}

\item{scale.u}{Prior scale (standard deviation) for the unpenalised
covariates.}

\item{regularized}{If \code{TRUE} (default), the regularized horseshoe prior
is used as opposed to the original horseshoe prior.}

\item{nu}{Number of degrees of freedom of the half-Student-t prior on the
local shrinkage parameters (by default, 1 if \code{regularized=TRUE}
and 3 otherwise).}

\item{par.ratio}{Expected ratio of non-zero to zero coefficients (ignored
if \code{regularized=FALSE}). The scale of the global shrinkage
parameter corresponds to \code{par.ratio} divided by the square root
of the number of observations; for linear regression only, it's further
multiplied by the residual standard deviation \code{sigma}.}

\item{global.df}{Number of degrees of freedom for the global shrinkage
parameter (ignored if \code{regularized=FALSE}).}

\item{slab.scale}{Scale of the regularization parameter (ignored if
\code{regularized=FALSE}).}

\item{slab.df}{Number of degrees of freedom of the regularization parameter
(ignored if \code{regularized=FALSE}).}

\item{qr}{Whether the QR decomposition should be used to decorrelate the
predictors (\code{TRUE} by default). This is silently set to
\code{FALSE} if there are more predictors than observations.}

\item{seed}{Optional integer defining the seed for the pseudo-random number
generator.}

\item{adapt.delta}{Target average proposal acceptance probability for
adaptation, a value between 0.8 and 1 (excluded). If unspecified,
it's set to 0.99 for hierarchical shrinkage models and to 0.95 for
base models.}

\item{...}{Further arguments passed to \code{\link[rstan]{sampling}},
such as \code{chains} (4 by default), \code{cores} (the value
of \code{options("mc.cores")} by default), \code{refresh}
(\code{iter / 10} by default).}
}
\description{
Run the No-U-Turn Sampler (NUTS) as implemented in Stan over all the data
or within the cross-validation folds.
}
