% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{lm_fprojsel}
\alias{lm_fprojsel}
\title{Forward selection minimizing KL-divergence in projection}
\usage{
lm_fprojsel(samples, max.num.pred = 30, out.csv = NULL)
}
\arguments{
\item{samples}{Object of class \code{stanfit}.}

\item{max.num.pred}{Maximum number of predictors after which the selection
procedure should stop.}

\item{out.csv}{If not \code{NULL}, the name of a CSV file to save the
output to.}
}
\description{
Forward selection minimizing KL-divergence in projection
}
