% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{lm_proj}
\alias{lm_proj}
\title{Compute projections of full predictors on to subspace of predictors}
\usage{
lm_proj(x, fit, sigma2, indproj, is.logistic)
}
\arguments{
\item{x}{Design matrix.}

\item{fit}{Matrix of fitted values for the full model.}

\item{sigma2}{Residual variance (1 for logistic regression).}

\item{indproj}{Vector of indices of the columns of \code{x} that form the
projection subspace.}

\item{is.logistic}{Set to \code{TRUE} for a logistic regression model.}
}
\description{
Compute projections of full predictors on to subspace of predictors
}
\keyword{internal}
