% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{log_lik.hsstan}
\alias{log_lik.hsstan}
\alias{log_lik}
\title{Pointwise log-likelihood}
\usage{
\method{log_lik}{hsstan}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{hsstan}.}

\item{newdata}{Optional data frame to use to evaluate the log-likelihood.
If \code{NULL} (default), the model matrix is used.}

\item{...}{Currently ignored.}
}
\value{
A matrix of size \code{S} by \code{N}, where \code{S} is number of draws from the posterior
distribution, and \code{N} is the number of data points.
}
\description{
Compute the pointwise log-likelihood.
}
\examples{
\dontshow{utils::example("hsstan", echo=FALSE)}
# continued from ?hsstan
log_lik(hs.biom)

}
