% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{loo.hsstan}
\alias{loo.hsstan}
\alias{loo}
\title{Approximate leave-one-out cross-validation}
\usage{
\method{loo}{hsstan}(x, save.psis = FALSE,
  cores = getOption("mc.cores"))
}
\arguments{
\item{x}{An object of class \code{hsstan}.}

\item{save.psis}{Whether intermediate results should be saved in the
returned object (\code{FALSE} by default).}

\item{cores}{Number of cores used for parallelisation (the value of
the \code{mc.cores} option by default).}
}
\value{
A \code{loo} object.
}
\description{
This computes an efficient approximate leave-one-out cross-validation
using Pareto smoothed importance sampling (PSIS-LOO).
}
