% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{nsamples.hsstan}
\alias{nsamples.hsstan}
\alias{nsamples}
\title{Number of posterior samples}
\usage{
\method{nsamples}{hsstan}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hsstan}.}

\item{...}{Currently ignored.}
}
\value{
The total number of posterior samples across the chains after discarding
the warmup iterations.
}
\description{
Extracts the number of posterior samples stored in a fitted model.
}
\examples{
\dontshow{utils::example("hsstan", echo=FALSE)}
# continued from ?hsstan
nsamples(hs.biom)

}
