% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{plot.projsel}
\alias{plot.projsel}
\title{Plot of the incremental contribution of each predictor}
\usage{
\method{plot}{projsel}(x, title = NULL, max.labels = NULL,
  font.size = 12, hadj = 0.05, vadj = 0.02, ...)
}
\arguments{
\item{x}{A data frame created by \code{\link{projsel}}.}

\item{title}{Title of the plot. If \code{NULL}, no title is displayed.}

\item{max.labels}{Maximum number of points to be labelled. If \code{NULL},
all those present in the \var{sel} file are displayed.}

\item{font.size}{Font size used to scale all text in the plot.}

\item{hadj, vadj}{Horizontal and vertical adjustment for the labels.}

\item{...}{Other options to plot() (currently ignored).}
}
\value{
A \pkg{ggplot2} object showing the relative incremental contribution of each
predictor starting from the initial set of unpenalized covariates.
}
\description{
A function of name \code{getfullname} to match variable names to full
names is searched on the current workspace, and if found it is used to
transform the labels before creating the plot.
}
