% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{posterior.means}
\alias{posterior.means}
\title{Returns the posterior means for the specified variables}
\usage{
posterior.means(samples, varnames)
}
\arguments{
\item{samples}{An object of class \code{stanfit}.}

\item{varnames}{Vector of variable names to be extracted.}
}
\description{
Returns the posterior means for the specified variables
}
