% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{print.hsstan}
\alias{print.hsstan}
\title{Print a summary for the fitted model}
\usage{
\method{print}{hsstan}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hsstan}.}

\item{...}{Further arguments to \code{\link[=summary]{summary()}}.}
}
\description{
Print a summary for the fitted model
}
