% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{sample.stan.cv}
\alias{sample.stan.cv}
\title{Runs a cross-validated Stan model}
\usage{
sample.stan.cv(stan.file, x, y, covariates, biomarkers, folds,
  standardize = TRUE, store.samples = TRUE, nu = 3, model.type = c("mc",
  "vb"))
}
\arguments{
\item{stan.file}{Path to a Stan model file.}

\item{x}{Data.frame of predictors.}

\item{y}{Vector of outcomes. For a logistic regression model, this is
expected to contain only \code{0-1} entries.}

\item{covariates}{Names of the variables to be used as (unpenalized)
predictors.}

\item{biomarkers}{Names of the variables to be used as penalized predictors.
If it is specified as an empty vector, a model with only unpenalized
covariates is used.}

\item{folds}{List of cross-validation folds, where each element contains
the indices of the test observations.}

\item{standardize}{Whether the design matrix should be standardized
(\code{TRUE} by default).}

\item{store.samples}{Whether the posterior samples should be saved
(\code{TRUE} by default).}

\item{nu}{Number of degrees of freedom for the half-Student-t priors.}

\item{model.type}{Either \code{"mc"} for Hamiltonian Monte Carlo, or
\code{"vb"} for variational Bayes.}
}
\description{
Runs either with Hamiltonian Monte Carlo or variational Bayes over the
cross-validation folds.
}
