% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{sample.stan}
\alias{sample.stan}
\alias{sample.stan.cv}
\title{Deprecated functions to fit hierarchical shrinkage models}
\usage{
sample.stan(x, y, covariates, biomarkers = NULL, logit = FALSE, ...)

sample.stan.cv(x, y, covariates, biomarkers = NULL, folds,
  logit = FALSE, iter = 1000, ...)
}
\arguments{
\item{x}{Data frame of predictors.}

\item{y}{Vector of outcomes. For a logistic regression model, this is
expected to contain only \code{0-1} entries.}

\item{covariates}{Names of the variables to be used as (unpenalized)
predictors.}

\item{biomarkers}{Names of the variables to be used as penalized predictors.
If it is specified as an empty vector, a model with only unpenalized
covariates is used.}

\item{logit}{\code{FALSE} for linear regression (default), \code{TRUE} for
logistic regression.}

\item{...}{Further options passed to \code{hsstan}.}

\item{folds}{Integer vector with one element per observation indicating the
cross-validation fold in which the observation should be withdrawn.
If \code{NULL} (default), no cross-validation is performed.}

\item{iter}{Total number of iterations in each chain, including warmup
(by default, 1000 iterations for cross-validation and 2000 otherwise).}
}
\description{
Deprecated functions to fit hierarchical shrinkage models
}
\section{Note}{

\code{sample.stan} and \code{sample.stan.cv} are thin wrappers around
\code{hsstan} provided for backward compatibility. They are considered
deprecated.
}

