% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{sampler.stats}
\alias{sampler.stats}
\title{Sampler statistics}
\usage{
sampler.stats(object)
}
\arguments{
\item{object}{An object of class \code{hsstan}.}
}
\value{
A matrix with \code{C + 1} rows, where \code{C} is the number of Markov
chains, reporting average acceptance probability, average stepsize, number
of divergent transitions, maximum tree depth, total number of gradient
evaluations, warmup and sampling times in seconds.
}
\description{
Report statistics on the parameters used in the sampler, the sampler
behaviour and the sampling time.
}
\examples{
\dontshow{utils::example("hsstan", echo=FALSE)}
# continued from ?hsstan
sampler.stats(hs.biom)

}
