% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{summarize.params}
\alias{summarize.params}
\title{Returns posterior means and 95\% credible intervals}
\usage{
summarize.params(samples, pars, varnames)
}
\arguments{
\item{samples}{An object of class \code{stanfit}.}

\item{pars}{Vector of variables names to be extracted.}

\item{varnames}{Vector of variable names.}
}
\description{
Returns posterior means and 95\% credible intervals
}
