% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{get.coefficients}
\alias{get.coefficients}
\title{Returns the posterior means of the regression coefficients}
\usage{
get.coefficients(samples, coeff.names)
}
\arguments{
\item{samples}{An object of class \code{stanfit}.}

\item{coeff.names}{Vector of names for the coefficients.}
}
\description{
Returns the posterior means of the regression coefficients
}
