% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{get.cv.performance}
\alias{get.cv.performance}
\title{Performance measures}
\usage{
get.cv.performance(obj, out.csv = NULL)
}
\arguments{
\item{obj}{An object or a list of objects of class \code{hsstan}.}

\item{out.csv}{Optional name of a file where the output can be saved in
CSV format.}
}
\description{
Extract measures of performance from plain or cross-validated results.
}
