% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{hsstan}
\alias{hsstan}
\title{Fit a plain or cross-validated Stan model}
\usage{
hsstan(x, y, covariates, biomarkers = NULL, family = gaussian,
  folds = NULL, iter = ifelse(is.null(folds), 2000, 1000),
  warmup = iter/2, scale.u = 2, regularized = TRUE,
  nu = ifelse(regularized, 1, 3), par.ratio = 0.05, global.df = 1,
  slab.scale = 2, slab.df = 4, store.samples = is.null(folds),
  seed = 123, adapt.delta = NULL, model.type = c("mc", "vb"))
}
\arguments{
\item{x}{Data frame of predictors.}

\item{y}{Vector of outcomes. For a logistic regression model, this is
expected to contain only \code{0-1} entries.}

\item{covariates}{Names of the variables to be used as (unpenalized)
predictors.}

\item{biomarkers}{Names of the variables to be used as penalized predictors.
If it is specified as an empty vector, a model with only unpenalized
covariates is used.}

\item{family}{Type of model fitted: either \code{gaussian()} for linear
regression (default) or \code{binomial()} for logistic regression.}

\item{folds}{List of cross-validation folds, where each element contains
the indices of the test observations. If \code{NULL} (default), no
cross-validation is performed.}

\item{iter}{Total number of iterations in each chain, including warmup
(by default, 1000 iterations for cross-validation and 2000 otherwise).}

\item{warmup}{Number of warmup iterations per chain (by default, half the
total number of iterations).}

\item{scale.u}{Prior scale (standard deviation) for the unpenalised
covariates.}

\item{regularized}{If \code{TRUE} (default), the regularized horseshoe prior
is used as opposed to the original horseshoe prior.}

\item{nu}{Number of degrees of freedom of the half-Student-t prior on the
local shrinkage parameters (by default, 1 if \code{regularized=TRUE}
and 3 otherwise).}

\item{par.ratio}{Expected ratio of non-zero to zero coefficients (ignored
if \code{regularized=FALSE}). The scale of the global shrinkage
parameter corresponds to \code{par.ratio} divided by the square root
of the number of observations; for linear regression only, it's further
multiplied by the residual standard deviation \code{sigma}.}

\item{global.df}{Number of degrees of freedom for the global shrinkage
parameter (ignored if \code{regularized=FALSE}).}

\item{slab.scale}{Scale of the regularization parameter (ignored if
\code{regularized=FALSE}).}

\item{slab.df}{Number of degrees of freedom of the regularization parameter
(ignored if \code{regularized=FALSE}).}

\item{store.samples}{Whether the posterior samples should be saved
(by default, \code{FALSE} for cross-validation and \code{TRUE}
otherwise).}

\item{seed}{Integer defining the seed for the pseudo-random number generator.}

\item{adapt.delta}{Target average proposal acceptance probability for
adaptation, a value between 0.8 and 1 (excluded). If unspecified,
it's set to 0.99 for hierarchical shrinkage models and to 0.95 for
base models.}

\item{model.type}{Either \code{"mc"} for Hamiltonian Monte Carlo, or
\code{"vb"} for variational Bayes.}
}
\description{
Runs either with Hamiltonian Monte Carlo or variational Bayes over the
cross-validation folds (if specified) or over all the data.
}
\details{
The continuous predictors and outcome variable are standardized before
fitting the models. In case of cross-validation, the test fold is
standardized according to mean and standard deviation of the corresponding
training fold.
}
