% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{plot.projsel}
\alias{plot.projsel}
\title{Plot of the incremental contribution of each predictor}
\usage{
plot.projsel(sel, title, filename = NULL, max.labels = NULL,
  font.size = 12, width = 800, height = 800)
}
\arguments{
\item{sel}{A data.frame created by \code{\link{lm_fprojsel}}.}

\item{title}{Title of the plot.}

\item{filename}{Optional name of a file where the plot is saved in png
format. If \code{NULL}, no png file is produced.}

\item{max.labels}{Maximum number of points to be labelled. If \code{NULL},
all those present in the \var{sel} file are displayed.}

\item{font.size}{Font size used to scale all text in the plot.}

\item{width}{Width of the png file in pixels.}

\item{height}{Height of the png file in pixels.}
}
\description{
A function of name \code{"getfullname"} to match variable names to full
names is searched on the current workspace, and if found it is used to
transform the labels before creating the plot.
}
