% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan.R
\name{sample.stan}
\alias{sample.stan}
\title{Runs a Stan model on all available data.}
\usage{
sample.stan(x, y, covariates, biomarkers = NULL, logit = FALSE,
  standardize = TRUE, adapt.delta = 0.99, iter = 2000,
  warmup = iter/2, scale.u = 20, nu = 3, model.type = c("mc",
  "vb"))
}
\arguments{
\item{x}{Data.frame of predictors.}

\item{y}{Vector of outcomes. For a logistic regression model, this is
expected to contain only \code{0-1} entries.}

\item{covariates}{Names of the variables to be used as (unpenalized)
predictors.}

\item{biomarkers}{Names of the variables to be used as penalized predictors.
If it is specified as an empty vector, a model with only unpenalized
covariates is used.}

\item{logit}{\code{FALSE} for linear regression (default), \code{TRUE} for
logistic regression.}

\item{standardize}{Whether the design matrix should be standardized.}

\item{adapt.delta}{Target average proposal acceptance probability for
adaptation (0.99 by default).}

\item{iter}{Number of iterations in each chain (including warmup).}

\item{warmup}{Number of warmup iterations per chain (by default, half the
total number of iterations).}

\item{scale.u}{Prior scale (standard deviation) for the unpenalised
covariates.}

\item{nu}{Number of degrees of freedom for the half-Student-t priors.}

\item{model.type}{Either \code{"mc"} for Hamiltonian Monte Carlo, or
\code{"vb"} for variational Bayes.}
}
\description{
Runs a Stan model on all available data.
}
